requirements.txt
# Core Dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx>=0.24.0

# AI/ML
huggingface-hub>=0.23.0

# Development (optional)
pytest==8.1.1
black==24.2.0
isort==5.13.2
